Move DesktopStoragePathsTest to core/platform module
The test was in feature/homescreen but tests core.platform.app code.
Also removed redundant KDoc from DesktopStoragePaths.
